content/canvas/test/test_2d.shadow.attributes.shadowColor.1.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.shadow.attributes.shadowColor.1</title>
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.shadowColor = 'lime';
ok(ctx.shadowColor === '#00ff00', "ctx.shadowColor === '#00ff00'");
ctx.shadowColor = 'RGBA(0,255, 0,0)';
ok(ctx.shadowColor === 'rgba(0, 255, 0, 0.0)', "ctx.shadowColor === 'rgba(0, 255, 0, 0.0)'");

SimpleTest.finish();

});
</script>
